Porting to Git Bash shell script
This variant is for use with Git Bash for Windows, including the portable variant; to perform the same function.  However, because its written for bash, and therefore gives me a bit more flexibility in how I can do things, I've changed a few things around.

### Command-Line Options
Thanks to the fancy argument parsing compared to batch; I've implemented the following:
```bash
-h, --help                                      List Usage for script
-f, --force                                      Do not prompt the user to overwrite files or create directories
-b, --backup                                  Perform a backup operation
-r, --restore                                   Perform a restore operation
-v, --verbose                                 Enable Verbose Logging
-t, --trim                                       Trim trailing slashes -- this shouldn't be necessary, though
-p, --profile <name>                     Backups/Restores placed into profile directories; specify the name here
--start-menu                                 Include the Start Menu in the backup/restore operation
--task-bar                                     Include the Task Bar in the backup/restore operation
--both                                           Include both the Task Bar and Start Menu in the backup/restore operation; overrides --start-menu and --task-bar
--root-dir <path>                         Specify the path to the root of the script storage; default: $HOME/Desktop/STB
```

### Variables
I've tried to split as much as I can into variables, which are all configurable within Lines 380-416.  Everything from the paths to each of the profile's contents location, to the command-line options for `cp`.

### Error Checking
Error checking in Batch is always really hard for me to deal with in any sensible way--and while it still isn't totally sensible here, its at least checking quite a bit more:
* Null values for vars, where that will cause a problem
* Vars meant to be set to 0 or 1 will be normalized to 0 if set to an invalid value
* Script storage paths will be generated if they don't already exist and the user provides permission
* Checks for invalid system paths
* Checks for nonexistent registry keys

I'll work on porting some of these changes to the batch variant, if I feel so inclined....or if I get bored, whichever comes first.